backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
PyMySQL==1.1.0
python-decouple==3.8
reportlab==4.0.7
openpyxl==3.1.2
pandas
django-filter==23.5
celery==5.3.4
redis==5.0.1
gunicorn
mysqlclient

